refactor: update pages router structure
