+++ src/pages/detail/index.jsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom'
-import useUserData from '../hooks/useUserData'
+import useUserData from '../../hooks/useUserData'
 
 const Detail = () => {
   const { data, isLoading, isError } = useUserData()
